requirement.txt
sanic~=23.3.0
sanic-ext~=23.3.0

flake8~=6.0.0
fabric~=3.1.0
greenlet>=3.0.0
SQLAlchemy~=2.0.16
werkzeug~=2.3.6
aiosqlite~=0.19.0
sanic-jwt~=1.8.0
filetype~=1.2.0
rarfile~=4.0
psutil~=5.9.5
Cython~=3.0.0
asyncssh~=2.13.2

jinja2
progressbar33
ruamel_yaml
pycryptodome
python-magic
